# Simplify expected errors function