The generic unrolled implementation performs similar to the SSE2 tuned code. This should make it faster on ARM platforms as well. In addition it is much easier to read.

```
# no filter (for reference)
Benchmark 1: cutadapt -o /dev/null  ~/test/5millionreads_R1.fastq
  Time (mean ± σ):      4.958 s ±  0.135 s    [User: 4.554 s, System: 0.399 s]
  Range (min … max):    4.786 s …  5.204 s    10 runs

# Before
Benchmark 1: cutadapt -o /dev/null --max-aer 0.01 ~/test/5millionreads_R1.fastq
  Time (mean ± σ):      6.608 s ±  0.092 s    [User: 6.192 s, System: 0.410 s]
  Range (min … max):    6.490 s …  6.800 s    10 runs

# After
Benchmark 1: cutadapt -o /dev/null --max-aer 0.01 ~/test/5millionreads_R1.fastq
  Time (mean ± σ):      6.612 s ±  0.132 s    [User: 6.197 s, System: 0.409 s]
  Range (min … max):    6.357 s …  6.881 s    10 runs 
```

Just some light hearted code simplifcation to make my day slightly more interesting (currently meddling with galaxy deployment and I ran cutadapt as a test, which reminded me of this code).