@rhpvorderman 2024-07-23
For reference, here is the performance for just the scalar loop:
```
Benchmark 1: cutadapt -o /dev/null --max-aer 0.01 ~/test/5millionreads_R1.fastq
  Time (mean ± σ):      7.113 s ±  0.157 s    [User: 6.690 s, System: 0.418 s]
  Range (min … max):    6.869 s …  7.434 s    10 runs
```
So the unrolling really helps.