@marcelm 2024-09-13
Indeed a nice improvement in readability! Awesome that it’s possible to trust the compiler here.