@rhpvorderman 2024-09-13
Another advantage is that the computation should generate the same result regardless of architecture, which is not guaranteed if specific intrinsics are used. 
I learned this from feedback at a PR at htslib. Great folks all around!